requirements.txt
click==5.1
coverage==4.0.1
decorator==4.0.4
funcsigs==0.4
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
mock==1.3.0
path.py==8.1.2
pbr==1.8.1
pep8==1.6.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
praw==3.3.0
ptyprocess==0.5
py==1.4.30
pytest==2.8.2
pytest-cov==2.2.0
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
update-checker==0.11
wheel==0.24.0
